requirements.txt
astroid==2.3.3
baidu-aip==2.2.18.0
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
Django==2.2.7
django-debug-toolbar==2.1
django-mysql==3.2.0
django-shortuuidfield==0.1.3
django-simpleui==2.9
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.4.1
matplotlib==3.1.2
mccabe==0.6.1
mysqlclient==1.4.4
numpy==1.17.4
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
Pillow==6.2.1
psycopg2==2.8.4
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.5
pytesseract==0.3.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.2
redis==3.3.11
requests==2.22.0
selenium==3.141.0
shortuuid==0.5.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
